requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
paho-mqtt>=1.5.0
requests>=2.19.1
aiohttp[speedups]>=3.7.4.post0
